made separate component folders and made a table
