+++ leeterboard/src/components/Board/Board.css
@@ -1,0 +1,4 @@
+.row {
+  /* position: relative;
+  display: flex; */
+}
